requirements.txt
torch>=2.2.0
torchvision>=0.17.0
opacus>=1.5.2
numpy>=1.26.0
pandas>=2.2.0
matplotlib>=3.8.0
scipy>=1.12.0
pyyaml>=6.0.1
tqdm>=4.66.0
